Added ABC integration
Provided function tune_hyperparameters which will create an instance of the artificial bee colony, run it, and then apply the values found to the ecnet model. 

Additionally stored a script to run the neural network in a function, and value ranges in a variable to easily call them to the artificial bee colonies' constructor.